refactor: move channel subscription gate into mcpserver package
Relocate the Channel watchset, the watch_task/unwatch_task/
list_watched_tasks tools, and the mute-by-default forwarding gate from
internal/mcpbridge into internal/server/mcpserver. Reuse the existing
jsonResult helper and rename the channel instructions constant to
ChannelInstructions to avoid colliding with mcpserver.Instructions.
command/mcp.go wires up mcpserver.NewChannel instead.
