+++ internal/server/mcpserver/channel_test.go
@@ -1,4 +1,4 @@
-package mcpbridge
+package mcpserver
 
 import (
 	"context"
